begins restructuring: works and is better, but not finished
