+++ app/staticIntentAnswering/queryBuilder.js
